backend/requirements.txt
django
djangorestframework
python-dotenv
django-extensions
django-filter
sqlparse
psycopg2